updated routes 13feb
